authentication/requirements/base.txt
Django
djangorestframework
django-filter
drf-spectacular
drf-spectacular-sidecar
drf-standardized-errors
# encrypted-model-fields
whitenoise
dj-database-url
djangorestframework-simplejwt
psycopg2-binary
django-redis
django-debug-toolbar
django-cors-headers
pika
django-redis